+++ core/src/docker.rs
@@ -1,1 +1,0 @@
-// TODO: migrate from src-tauri and agent