+++ core/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod config;
-pub mod docker;
 pub mod errors;
 pub mod files;
 pub mod monitoring;